Initial RPM build
This build an RPM for SUSE/RedHat that installs the `sh/` library into
`/usr/lib/csm-common-library`. The RPM version is determined by the
return from `git describe --tags`, the return will determine if the
built RPM is stable or unstable. "Stable" RPMs are created when a semver
git-tag is taken (vX.Y.Z), "Unstable" RPMs will be created by any commit
that does not have a semver tag.